### 2024-04-23 #6666 @editorialbot
[PRE REVIEW]: superblockify: A Python Package for Automated Generation, Visualization, and Analysis of Potential Superblocks in Cities
**Submitting author:** @cbueth (<a href="http://orcid.org/0000-0003-2298-8438">Carlson Büth</a>)
**Repository:** https://github.com/NERDSITU/superblockify
**Branch with paper.md** (empty if default branch): submission
**Version:** 1.0.0
