### 2024-04-23 ee498ee
Update bibliographies, add dois and urls
