+++ paper/paper.bib
@@ -4,7 +4,8 @@
   journal={bmj},
   volume={372},
   year={2021},
-  publisher={British Medical Journal Publishing Group}
+  publisher={British Medical Journal Publishing Group},
+  doi={10.1136/bmj.n443}
 }
 
 @article{miner2024car,
@@ -14,7 +15,8 @@
   volume={115},
   pages={103817},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jtrangeo.2024.103817}
 }
 
 @article{mattioli_political_2020,
@@ -82,12 +84,6 @@ note={(accessed March 5, 2024)}
   url          = {https://doi.org/10.5281/zenodo.10476253}
 }
 
-@misc{qgis,
-title = {QGIS},
-year = {2024},
-howpublished = "\url{https://qgis.org/}",
-note={(accessed March 5, 2024)}}
-
 @Manual{qgis,
   title = {QGIS Geographic Information System},
   author = {{QGIS Development Team}},
@@ -133,7 +129,8 @@ note={(accessed March 5, 2024)}}
 @techreport{frey2020potenziale,
   title={Potenziale von Superblock-Konzepten als Beitrag zur Planung energieeffizienter Stadtquartiere-SUPERBE},
   author={Frey, Harald and Leth, Ulrich and Sandholzer, Fabian Josef},
-  year={2020}
+  year={2020},
+  url={https://nachhaltigwirtschaften.at/en/sdz/projects/superbe.php}
 }
 
 @article{eggimann_potential_2022,
@@ -180,7 +177,8 @@ Publisher: Nature Publishing Group},
   title={Walking and cycling: latest evidence to support policy-making and practice},
   author={WHO},
   year={2022},
-  publisher={World Health Organization. Regional Office for Europe}
+  publisher={World Health Organization. Regional Office for Europe},
+  url={https://apps.who.int/iris/handle/10665/354589}
 }
 
 @article{nieuwenhuijsen_superblock_2024,
@@ -352,7 +350,8 @@ Publisher: Nature Publishing Group},
   date = {2021},
   series = {Ph.d.-Serien for {{Det Tekniske Fakultet}} for {{IT}} Og {{Design}}, {{Aalborg Universitet}}},
   publisher = {{Aalborg Universitetsforlag}},
-  url = {https://vbn.aau.dk/en/publications/points-of-exchange-spatial-strategies-for-the-transition-towards-}
+  url = {https://vbn.aau.dk/en/publications/points-of-exchange-spatial-strategies-for-the-transition-towards-},
+  doi = {10.54337/aau451017237}
 }
 
 @article{mueller2020,
