### 2024-04-23 #6666 @cbueth
Thank you for starting the pre review step.

We consider the following reviewers to have a suitable background and suggest them:

- lukasbeuster
- jkbren
- dinacmistry
- LCB0B
- jGaboardi
- tomalrussell

Also, we checked all bib entries without DOI and added them where possible, alternatively a url to the project page, to the updated submission branch. See commit https://github.com/NERDSITU/superblockify/commit/ee498eec20a48314236c088ed6ee465bc10ee4b7.
