### 2024-04-23 fa3c695
Fix: Set date to today
