### 2024-05-31 NERDSITU/superblockify#86 @erexer
No module named 'scipy' error
Hi, I'm a reviewer for JOSS, excited to learn more about the package!
https://github.com/openjournals/joss-reviews/issues/6798

started a virtual env for `python3.10.14`, ran
`$ pip install superblockify`
`$ python examples/00-mwe.py`

it threw
```
$ python examples/00-mwe.py
2024-05-31 15:37:43,244 |  WARNING | font_manager.py:1037 | Matplotlib is building the font cache; this may take a moment.
2024-05-31 15:38:00,135 |     INFO | font_manager.py:1049 | Failed to extract font properties from /System/Library/Fonts/LastResort.otf: tuple indices must be integers or slices, not str
2024-05-31 15:38:00,149 |     INFO | font_manager.py:1049 | Failed to extract font properties from /System/Library/Fonts/Supplemental/NISC18030.ttf: In FT2Font: Could not set the fontsize (invalid pixel size; error code 0x17)
2024-05-31 15:38:00,685 |     INFO | font_manager.py:1049 | Failed to extract font properties from /System/Library/Fonts/Apple Color Emoji.ttc: In FT2Font: Could not set the fontsize (invalid pixel size; error code 0x17)
2024-05-31 15:38:00,778 |     INFO | font_manager.py:1581 | generated new fontManager
Traceback (most recent call last):
  File "<mypath>/superblockify/examples/00-mwe.py", line 7, in <module>
    from superblockify import ResidentialPartitioner
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/__init__.py", line 7, in <module>
    from ._api import *
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/_api.py", line 5, in <module>
    from .partitioning import BearingPartitioner
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/partitioning/__init__.py", line 6, in <module>
    from .base import BasePartitioner
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/partitioning/base.py", line 20, in <module>
    from .checks import is_valid_partitioning
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/partitioning/checks.py", line 12, in <module>
    from ..utils import has_pairwise_overlap
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/utils.py", line 28, in <module>
    from .partitioning.utils import reduce_graph
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/partitioning/utils.py", line 18, in <module>
    from ..graph_stats import get_population_area, basic_graph_stats
  File "<mypath>/superblockify/.venv/lib/python3.10/site-packages/superblockify/graph_stats.py", line 7, in <module>
    from scipy.stats import entropy
ModuleNotFoundError: No module named 'scipy'
```

would recommend adding `scipy` to the list of dependencies since its used by `graph_stats.py`
