### 2024-06-05 a3ecd3f
Correct url in CONTRIBUTING.md
