### 2024-06-11 NERDSITU/superblockify#87 @erexer
JOSS Review: Documentation 
This is part of your [JOSS review ](https://github.com/openjournals/joss-reviews/issues/6798).

Documentation
- [x] “A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?” Can you add to the statement of need to specify who the target audience is.
- [x] Might be good to clarify that the `data/` dir will be created when the example is run (not required for JOSS acceptance).
- [x] I'm not finding a `Contribute` section in the docs. If it's already in there, can you point me towards it, and if not, can you add a `Contribute` section with instructions on how to 1) Contribute to the software, 2) Report issues or problems with the software, and 3) Seek support.
