### 2024-06-12 bbd1857
Add `scipy` dependency explicitly - remove unused `pyarrow` - 1.0.0rc9
