+++ docs/conf.py
@@ -24,7 +24,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = "superblockify"
 copyright = "2023-2024, superblockify developers"
 author = "superblockify developers"
-release = "1.0.0"
+release = "1.0.0rc9"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
