### 2024-06-12 9af06d0
Linted `E0601` and `E0606`
