### 2024-06-12 1aa799e
Add `CONTRIBUTING.md` reference in doc index page - reset `html_sidebars` for index page - remove GNOME/byzanz tool from doc
