+++ CONTRIBUTING.md
@@ -157,8 +157,7 @@ as [GitHub issues](https://github.com/NERDSITU/superblockify/issues).
 - You may want to **include screenshots and animated GIFs** which help you demonstrate
   the steps or point out the part which the suggestion is related to. You can
   use [this tool](https://www.cockos.com/licecap/) to record GIFs on macOS and Windows,
-  and [this tool](https://github.com/colinkeenan/silentcast)
-  or [this tool](https://github.com/GNOME/byzanz) on
+  and [this tool](https://github.com/colinkeenan/silentcast) on
   Linux. <!-- this should only be included if the project has a GUI -->
 - **Explain why this enhancement would be useful** to most superblockify users. You may
   also want to point out the other projects that solved it better and which could serve
