### 2024-06-12 NERDSITU/superblockify#86 @cbueth
Hi there,

Thanks for pointing out the missing `scipy` dependency. We've added it to our dependencies list in bbd18576271e7f9e657dd816312a82fdafbbdaa5 and the changes are now live on the submission branch and in the PyPI package `1.0.0rc9`.

Just a quick note - our package also relies on `osmnx`, which can't be installed with pip alone. To set up the environment correctly, please use the following command from the `README.md`/Installation Page:

```bash
conda create -n sb_env -c conda-forge python=3.12 osmnx=1.9.2
```

Other than `python` and `osmnx` it also installs `scipy`. But as we use it directly, it is better to include it explicitly in the package dependencies. Afterwards you can activate the environment and use pip to install `superblockify`. Hopefully, this should sort out the issue you've been facing. Let us know!

Thanks again for your feedback. Let us know if you have any other issues or suggestions.
