### 2024-06-12 6a4cc74
Add note about automatic `data/` folder creation
