### 2024-06-12 32666ed
Set version to `1.0.0`
