+++ superblockify/_version.py
@@ -1,3 +1,3 @@
 """superblockify package version."""
 
-__version__ = "1.0.0rc9"
+__version__ = "1.0.0"
