### 2024-06-12 NERDSITU/superblockify#87 @cbueth
Thank you for your feedback. We have made and [deployed](https://superblockify.city/) the following changes based on your suggestions:

- a3998ff Thanks for the reminder. We have added a section with a shortened version of the Stetement of Need from the `paper.md`.

- 6a4cc74 We have added a note in the `examples/00-mwe.*` script and notebook to clarify that the `data/` directory will be automatically created when the examples are run.

- 1aa799e We have added a Contribute section in the documentation with a link to the `CONTRIBUTING.md`.

We hope these changes address your concerns. Please let us know if there's anything else we can improve.
