### 2024-06-12 #6798 @cbueth
Thank you all for the activity and she short reminder!

We are happy to see the many filled check marks and have just replied with the asked for changes in issues https://github.com/NERDSITU/superblockify/issues/86 and https://github.com/NERDSITU/superblockify/issues/87. The new release candidate `1.0.0rc9` is up on PyPI.

Please let us know if there are any further questions!
