### 2024-06-12 NERDSITU/superblockify#87 @erexer
Great! In the [`Contributing.md`](https://github.com/NERDSITU/superblockify/blob/main/CONTRIBUTING.md#suggesting-enhancements) file, the Table of Contents lists some sections that don't exist, can you remove those.

Do you welcome PR's from folks? If you do, I'd recommend adding that to the [`Contributing.md`](https://github.com/NERDSITU/superblockify/blob/main/CONTRIBUTING.md#suggesting-enhancements) file (not required for JOSS acceptance).
