### 2024-06-14 NERDSITU/superblockify#88 @erexer
JOSS review: questions on 00-mwe.py
This is for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6798).

I have a couple questions about the minimum working example. I'm running `00-mwe.py` after installing using micromamba and pip. I'm on a Mac v14.5, python v3.12.3.
1. Is "residential" referring to number of residential streets? If so, why does it have non-whole numbers? The y-axis should specify which "length" unit it's referring to.

2. in the x-axis label on this figure (Distribution of shortest path lengths on full graph) and for the "Distribution of shortest path distances for the partitioning" figure, what does "[s]" mean? Do the "Coordinates of nodes" graph x- and y-axis have units (lat/lon)?

3. I'm having an issue where the figures after "Distribution of shortest path distances for the partitioning" don't load if I wait 50 seconds or longer before I close the python window for "Distribution of shortest path distances for the partitioning", it pops up three python windows labeled "Figure 1", "Figure 2", and "Figure 3" which are all completely blank. When I hover over the window, it shows the spinny wheel of death and I have to force quit python. When I wait less than 50 seconds before closing "Distribution of shortest path distances for the partitioning", it pops up Figure 2, 3, and 4 fine and I can see the charts.

4. What is "Part." short for? Not sure abbreviating is necessary. Is "(s)" seconds?

5. I assume these are separate color bars since the left one says that dark green is 1 and the right color bar says that dark green is 400. Can you make it clear which graphs they apply to? If they aren't separate color bars, can you explain the scale? Again, what does "(s)" stand for?

6. Is there a glossary of terms somewhere?

If any of these are explained in the documentation, I would love to be pointed towards it! I may have missed it :)
