### 2024-06-14 NERDSITU/superblockify#86 @erexer
Thanks for the clarification! It might be helpful to point out the `osmnx` installation caveat in the installation guide since others might do what I did and prefer/use a different venv setup.
