### 2024-06-17 671f22a
Add note about installation C caveat - Installation with plain pip is not possible
