### 2024-06-17 NERDSITU/superblockify#86 @cbueth
Good point! Added a note to the installation instructions, see 671f22a.
