### 2024-07-04 #6798 @caimeng2
> 👋 @caimeng2 - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Thanks for the invite! I can review the package in the week of July 22. Will that work?
