### 2024-07-22 ee6f76f
Change plotting modularity - Show all plots after running - Better as `plt.show()` is a blocking function. Many windows, but program not stalling
