### 2024-07-22 faf9c1f
Improve plot labels - move distance matrix plot to logging mode 'debug'
