+++ superblockify/metrics/metric.py
@@ -20,7 +20,6 @@ from .measures import (
 )
 from .plot import (
     plot_distance_matrices,
-    plot_distance_matrices_pairwise_relative_difference,
     plot_component_wise_travel_increase,
     plot_relative_difference,
     plot_relative_increase_on_graph,
@@ -373,32 +372,23 @@ class Metric:
         partitioner : BasePartitioner
             The partitioner object to calculate the metrics for
         """
-
-        fig, _ = plot_distance_matrices(
-            self, name=f"{partitioner.name} - {partitioner.__class__.__name__}"
-        )
-        save_plot(
-            partitioner.results_dir,
-            fig,
-            f"{partitioner.name}_distance_matrices.{Config.PLOT_SUFFIX}",
-        )
-        fig, _ = plot_distance_matrices_pairwise_relative_difference(
-            self, name=f"{partitioner.name} - {partitioner.__class__.__name__}"
-        )
-        save_plot(
-            partitioner.results_dir,
-            fig,
-            f"{partitioner.name}_distance_matrices_"
-            f"pairwise_relative_difference.{Config.PLOT_SUFFIX}",
-        )
-        fig, _ = plot_relative_difference(
-            self, "N", "S", title=f"{partitioner.name} - {self.__class__.__name__}"
-        )
-        save_plot(
-            partitioner.results_dir,
-            fig,
-            f"{partitioner.name}_relative_difference_SN.{Config.PLOT_SUFFIX}",
-        )
+        if logger.getEffectiveLevel() <= 10:
+            fig, _ = plot_distance_matrices(
+                self, name=f"{partitioner.name} - {partitioner.__class__.__name__}"
+            )
+            save_plot(
+                partitioner.results_dir,
+                fig,
+                f"{partitioner.name}_distance_matrices.{Config.PLOT_SUFFIX}",
+            )
+            fig, _ = plot_relative_difference(
+                self, "N", "S", title=f"{partitioner.name} - Relative difference in {self.__class__.__name__}"
+            )
+            save_plot(
+                partitioner.results_dir,
+                fig,
+                f"{partitioner.name}_relative_difference_SN.{Config.PLOT_SUFFIX}",
+            )
         fig, _ = plot_component_wise_travel_increase(
             partitioner,
             self.distance_matrix,
