+++ superblockify/partitioning/plot.py
@@ -211,10 +211,10 @@ def plot_component_rank_size(partitioner, measure):
         color="k",
         zorder=2,
     )
-    axe.set_xlabel("Component rank", fontsize=12)
-    axe.set_ylabel(f"Component size ({measure} [m])", fontsize=12)
+    axe.set_xlabel("Superblock rank", fontsize=12)
+    axe.set_ylabel(f"Superblock size ({measure} [m])", fontsize=12)
     axe.set_title(
-        f"Component size rank for {partitioner.name} with attribute "
+        f"Superblock size rank for {partitioner.name} with attribute "
         f"`{partitioner.attribute_label}`",
         fontsize=13,
     )
