### 2024-07-22 c1385a8
Fix code style issues with Black
