+++ superblockify/metrics/plot.py
@@ -49,6 +49,7 @@ def plot_distance_distributions(
     axe[1].set_ylabel(labels[1])
     plt.tight_layout()
 
+
 def plot_distance_matrices(metric, name=None):
     """Show the distance matrices for the network measures.
 
