### 2024-07-22 NERDSITU/superblockify#88 @cbueth
Thank you very much for trying out our minimum working example and the notes.

Sorry for the late response. For some reasons issue email notifications were deactivated, but they are active now on our side.

From your notes we have implemented changes in ee6f76fe2b884f197af4c6778c97a93a320021ed and faf9c1febbe04ecc6573483b8d0c35ac87441a4b. With PR #89 the changes will be merged.

- (3.) We have generally changed when the plots are being shown. Now all plots will be shown at the end of the run() method at once. We were not able to reproduce the behavior you described, but this should fix it and is more aligned with the best practice. Thank you for this valuable hint.
- (1.) We have removed the "Component size of"-plot as it is too similar to the size rank plot, now called "Superblock size rank for".  The plot before was like the size rank plot, comparing the distribution of street length sum in each Superblock, but was unordered in contrast to the size rank plot.
![Figure_2](https://github.com/user-attachments/assets/ab907ad9-96df-439e-8424-a57ed49fd64d)
- (2.) For the path length distribution plots we improved the labeling. Shown are the projected coordinates in meters for Lat/Lon.
![Figure_4](https://github.com/user-attachments/assets/8466b04c-414c-4300-9425-cd390deb3dec)
- (4.) "Part." is partition abbreviated. We call areas partitions internally that later are superblocks, but sometimes partitions need to fuse to one superblock out of topological reasons. Finally, we renamed the lables for all plots this to say superblock(s).
- (5.) The "pairwise relative difference"-plot fuses the "Distance matrix" and "Relative difference" plots. In the case of having more than two distance metrics one would be able to compare each distance metric pairwise in one plot. We now decided only to show the two latter ones, and only if the user activated debug logging, as the insight from these plots is not seen as big. Also, the labeling on these two plots is clearer, e.g. having only one colorbar.
- (2.,4.,5.) Yes, we use SI units, so "s" stands for seconds and "m" for meters, the unit always in brackets. This means that distance must be understood as travel time when "(s)" is given. This is due to the feature that one can change the unit distance is measured.
- (6.) There is no central glossary of terms until now, you did not miss it. The documentation pages, including Usage and Reference pages should introduce the package's background.

Please, let us know if you have any questions related :clap:
