### 2024-07-22 894a08f
Tests: change plot hiding
