### 2024-07-22 a9c526b
Plots: unify unit brackets
