### 2024-07-24 #6798 @caimeng2
> Hey @caimeng2 👋, thank you for taking the time to review our submission. We are looking forward to your feedback.
>
> Just to note, there have been minor changes addressed regarding the minimum working example script @erexer noticed in [NERDSITU/superblockify#88](https://github.com/NERDSITU/superblockify/issues/88), but they have not been integrated into the latest release candidate, yet.

Good to know. Thanks!
