### 2024-07-26 NERDSITU/superblockify#88 @erexer
Great! I'm running `00-mwe.py` on the `88-joss-review-questions-on-00-mwepy` branch, but I'm still seeing a "Component size of Scheveningen_test" plot, was that supposed to be removed in (1.)?
