### 2024-07-26 NERDSITU/superblockify#88 @cbueth
For these changes to appear you need to install the package in editable mode from the repository, different than the user documentation says. Have you done this?

You can create a separate environment like so (from the `CONTRIBUTING.md`):

```bash
conda create -n sb_env_dev -c conda-forge python=3.12 -f environment-dev.yml
conda activate sb_env_dev
pip install --no-build-isolation --no-deps -e .
```
