### 2024-07-31 NERDSITU/superblockify#88 @erexer
Thanks for making these changes and for the editable installation instructions! It's running well on my end.

> (2.,4.,5.) This means that distance must be understood as travel time when "(s)" is given.
Gotcha, in that case I'd recommend changing the label on the graph to "Travel time (s)" since using seconds for distance is unintuitive.

> (2.) For the path length distribution plots we improved the labeling. Shown are the projected coordinates in meters for Lat/Lon.
I've never seen lat/lon in meters before, how does that translate from degrees?
