### 2024-08-01 NERDSITU/superblockify#88 @cbueth
Great to hear!

> Gotcha, in that case I'd recommend changing the label on the graph to "Travel time (s)" since using seconds for distance is unintuitive.

We'll see on how to elegantly implement that change, it's definitely more concise that way!

> I've never seen lat/lon in meters before, how does that translate from degrees?

As these are projected coordinates, projected to the current CRS of the place, these are in meters. This way we can use Euclidean distances. These are commonly called y and x, but somewhat correlate to latitude and longitude. I'd change it to y and x.

We'll make these changes if you find the choices reasonable and afterwards push a new release candidate if that would close this issue.
