### 2024-08-01 NERDSITU/superblockify#88 @erexer
Yep, making those changes would close this issue.
