### 2024-08-02 NERDSITU/superblockify#90 @caimeng2
comments on paper
The paper is well-written and presents its ideas with good organization. I have the following questions and comments for improvement.

1. The paper defines a Superblock as "a set of adjacent urban blocks where vehicular through traffic is prevented or pacified, giving priority to people walking and cycling," however, how was this concept _**operationalized**_ in the package? Did you quantify or measure "adjacent", "urban", "vehicular through traffic is prevented or pacified", and "giving priority to people walking and cycling"?

2. Related to the previous question, is the package's delineation of Superblocks always accurate? By "accurate," I mean does the delineation always comply with the definition? What would be the limitations, if any, of this package?

3. It is not clear to me whether the usage of this package is to detect **_existing_** Superblocks, or identify an area that is suitable for _**future**_ Superblock planning. Or both?
    - The first use case seems to be for planning. It would be clearer if you could state your criteria for selecting a potential area and explain why you chose those criteria.
    - The second use case seems to be for existing Superblocks as you talk about the impacts. (Also the overall write-up sounds more like it is for existing Superblocks.) Could you be more specific about what the "valuable insights" are (line 51)?

4. Line 70-74. Under what circumstances should users choose each partitioner?

5. Figure 1.
     - Do the different colors mean anything? If no, it's better to make it clear in the legend.
     - What does "representative nodes" in the description mean? Representative of what? I don't get what the colored nodes are.

6. The analysis feature could be better explained. Urban planners who are not familiar with the graph theory probably won't understand "global efficiency," "directness," etc. A short description of what each of the metrics means would be helpful.

Reference: https://github.com/openjournals/joss-reviews/issues/6798
