### 2024-08-02 NERDSITU/superblockify#90 @cbueth
Hey Meng, thank you for your valuable feedback on our paper and the compliments. Reading through your notes, we think they will improve the paper. Your feedback will be integrated coming week and we will notify you here about the changes and notes.
