### 2024-08-02 NERDSITU/superblockify#88 @cbueth
We have just merged the changes after a successful CI and published the release candidate [`1.0.0rc10` on PyPI](https://pypi.org/project/superblockify/) including the changes.
Thank you again for reviewing and testing our package!
