### 2024-08-02 #6798 @cbueth
We just merged the feedback from https://github.com/NERDSITU/superblockify/issues/88 by @erexer in PR https://github.com/NERDSITU/superblockify/pull/89 and released a  [`1.0.0rc10`](https://pypi.org/project/superblockify/1.0.0rc10/) with the changes to PyPI.

Coming week we'll get into the comments on the paper noted by @caimeng2 in https://github.com/NERDSITU/superblockify/issues/90.

Thanks everyone for the collaboration on the review! :raised_hands:
