### 2024-08-02 NERDSITU/superblockify#91 @caimeng2
OMP warning
I'm testing version 1.0.0rc10 with MacOS 14.5. I installed the package following the installation instructions and it went fine. Running `import superblockify as sb` shows a warning

> OMP: Info #276: omp_set_nested routine deprecated, please use omp_set_max_active_levels instead.

The code still runs but better to resolve the warning.

Reference: https://github.com/openjournals/joss-reviews/issues/6798
