### 2024-08-02 NERDSITU/superblockify#92 @caimeng2
TypeError: BasePartitioner.__init__() got an unexpected keyword argument 'percentile'
Running
```
part = sb.BetweennessPartitioner(
    name="Scheveningen_test_betweenness", city_name="Scheveningen",
    search_str="Scheveningen, NL",
    percentile=85.0,  # aimed percentage of streets for Superblocks
    scaling="normal",  # type of scaling for the betweenness centrality:
    # "normal", "length", "linear"
    max_range=None  # maximum range of the betweenness centrality, in meters, e.g. 3500
)
```
led to
> TypeError                                 Traceback (most recent call last)
> Cell In[8], line 1
> ----> 1 part = sb.BetweennessPartitioner(
>       2     name="Scheveningen_test_betweenness", city_name="Scheveningen",
>       3     search_str="Scheveningen, NL",
>       4     percentile=85.0,  # aimed percentage of streets for Superblocks
>       5     scaling="normal",  # type of scaling for the betweenness centrality:
>       6     # "normal", "length", "linear"
>       7     max_range=None  # maximum range of the betweenness centrality, in meters, e.g. 3500
>       8 )
>
> TypeError: BasePartitioner.__init__() got an unexpected keyword argument 'percentile'

I did a quick search in your source code. `percentile` is not one of the parameters in the`BasePartitioner` class.

Reference: https://github.com/openjournals/joss-reviews/issues/6798
