### 2024-08-02 NERDSITU/superblockify#91 @cbueth
Thank you for reporting this issue. The warning `OMP: Info #276: omp_set_nested routine deprecated, please use omp_set_max_active_levels instead.` is related to the OpenMP library, which is used for parallel programming.
It is good to know the code runs fine on your end, while it seems weird, that this warning already happens when only loading the module, before executing any code. As this is the case and we do not directly address OMP setting `omp_set_nested`, it must be some dependency that we don't have control about.
`numba` seems to do this in a prominent functionality, see
https://github.com/numba/numba/blob/2001717f3321a5082c39c5787676320e699aed12/numba/np/ufunc/omppool.cpp#L228
In this case I do not see a way we can avoid this, as it is in a core functionality (thread launcher for `ufunc`´s).

If there is some further hint that points otherwise, let us know.
