### 2024-08-02 f9084f0
Doc: Correct usage of additional parameters in example notebook and usage page
