### 2024-08-02 NERDSITU/superblockify#92 @cbueth
Thank you for noticing this! This code is from the `examples/01-getting-started.ipynb`, right?

The three given options can be passed to the `run()` method of the `BetweennessPartitioner`.

```python
part = sb.BetweennessPartitioner(
    name="Scheveningen_test_betweenness", city_name="Scheveningen",
    search_str="Scheveningen, NL",
)
part.run(
    percentile=85.0,  # aimed percentage of streets for Superblocks
    scaling="normal",  # type of scaling for the betweenness centrality:
    # "normal", "length", "linear"
    max_range=None  # maximum range of the betweenness centrality, in meters, e.g. 3500
)
```

We corrected the usage in that file and in [`docs/usage.myst`](https://superblockify.city/usage/#the-superblocks-look-too-big-small-random-when-using-the-residentialpartitioner-why-is-that), see f9084f0. To our knowledge that are the only occurrences of this.
