### 2024-08-06 ab5c166
Paper: minor changes for issue #90
