+++ paper/paper.md
@@ -38,14 +38,15 @@ bibliography: paper.bib
 
 # Summary
 
-`superblockify` is a Python package for partitioning an urban street network into
-Superblock-like neighborhoods and for visualizing and analyzing the partition results.
+`superblockify` is a Python package designed to assist in planning future Superblock
+implementations by partitioning an urban street network into Superblock-like neighborhoods
+and providing tools for visualizing and analyzing these partition results.
 A Superblock is a set of adjacent urban blocks where vehicular through traffic is
 prevented or pacified, giving priority to people walking and
-cycling [@nieuwenhuijsen_superblock_2024]. The Superblock blueprints and descriptive
+cycling [@nieuwenhuijsen_superblock_2024]. The potential Superblock blueprints and descriptive
 statistics generated by `superblockify` can be used by urban planners as a first step in
-a data-driven planning pipeline, or by urban data scientists as an efficient
-computational method to evaluate Superblock partitions. The software is licensed under
+a data-driven planning pipeline for future urban transformations, or by urban data scientists as an efficient
+computational method to evaluate potential Superblock partitions. The software is licensed under
 AGPLv3 and is available at \url{https://superblockify.city}.
 
 # Statement of need
@@ -92,7 +93,7 @@ This quick feedback can reduce the time and resources required to manually plan
 Superblocks, which in turn can accelerate sustainable urban development.
 Second, `superblockify` enables researchers to conduct large-scale studies across
 multiple cities or regions, providing valuable insights into the potential impacts of
-Superblocks at a broader scale.
+Superblocks at a broader scale, e.g. travel time changes.
 In both cases, `superblockify` can help to identify best practices, algorithmic
 approaches, and strategies for Superblock implementation.
 
@@ -130,6 +131,9 @@ In its current version 1.0.0, `superblockify` comes with two partitioners:
 2. The betweenness approach uses the streets with high betweenness centrality for the
    decomposition.
 
+The choice between these two approaches depends on the data quality and the desired outcome.
+The residential approach is appropriate for using residential data, if available and accurate.
+The betweenness approach is a flow-based alternative.
 The resulting Superblocks can be exported in GeoPackage (`.gpkg`) format for further
 use.
 
@@ -145,13 +149,18 @@ Example Superblock configurations for two cities are shown in Fig.
 
 ## Analysis
 
-For analysis, the package calculates various graph metrics of the street network, such
-as global efficiency [@latora2001], directness [@szell2022], betweenness
-centrality [@brandes2008], spatial clustering and anisotropy of high betweenness
-centrality nodes [@kirkley2018], street orientation-order [@boeing2019], or average
-circuity [@boeing2019a].
+For analysis, the package calculates various graph metrics of the street network, including:
+
+- Global efficiency [@latora2001]: In the context of Superblocks, this measures how the overall ease of movement across the city might change after implementation.
+- Directness [@szell2022]: This indicates how Superblock implementation might affect the directness of routes, potentially increasing or decreasing detours.
+- Betweenness centrality [@brandes2008]: Identifies which streets might bear increased traffic load after Superblock implementation.
+- Spatial clustering and anisotropy of high betweenness centrality nodes [@kirkley2018]: Describes how clustered and non-uniformly distributed the expected traffic bottlenecks are.
+- Street orientation-order [@boeing2019]: Quantifies how grid-like each Superblock is.
+- Average circuity [@boeing2019a]: Measures the windiness of routes compared to straight-line distances.
+
 These metrics are calculated for the entire street network and for each Superblock
-individually.
+individually, providing insights into how the Superblock implementation might affect the
+overall city structure and local neighborhood characteristics.
 To facilitate further analysis, all of these metrics are included in the exportable
 GeoPackage file.
 
