### 2024-08-06 26f0ed0
Plots: Add headings and legend
