### 2024-08-06 3c5bdb9
Doc: Add GHSL maintenance check - We don't want to rebuild page if GHSL dataset would not be reachable
