### 2024-08-10 #6798 @crvernon
@cbueth - wow, this submission is very clean!  Thank you for making my job easy!  I have only one needed edit at this point:

- [ ] In the caption for Figure 1, you simply state the city names.  Could you also place these into their global context?  E.g., Baltimore, MD, USA.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software if you haven't already and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on GitHub.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the paper exactly and the order in which they are listed; make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

- [ ]  Please respond with the 1) DOI of the archived version and the 2) version here.
