### 2024-08-12 01892f1
Contextualize locations in paper.md
