### 2024-08-13 #6798 @cbueth
@crvernon Thank you for checking and editing our submission. We are delighted to hear for you that our submission is clean!

- https://github.com/NERDSITU/superblockify/commit/01892f179872d77f2884f875ee62609c970d3748: The caption has been updated with the more specific location strings:
  "Athens, GR (top row) and Baltimore, MD, USA (bottom row) [...]"
- We have [tagged and released 1.0.0](https://github.com/NERDSITU/superblockify/releases/tag/1.0.0). The same release has been [published on Zenodo](https://zenodo.org/records/13300611) with the filled in metadata, same title and license.
1. DOI for 1.0.0: [10.5281/zenodo.13300611](https://doi.org/10.5281/zenodo.13300611) (for all versions [10.5281/zenodo.13300610](https://zenodo.org/doi/10.5281/zenodo.13300610))
2. The version number for the release is as planned version `1.0.0`.

This should be all.
