### 2024-08-13 7335454
Add `CITATION.cff` & DOI
