### 2024-08-14 #6798 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
