### 2024-08-14 #6798 @crvernon
@cbueth I noticed that you have backticks around the title in your paper.  This causes formatting issues in the way our site formats your publication.  Could you remove these?   No need to conduct a new release or anything.

Let me know when this is done and I'll finalize this one.

Thanks.
