### 2024-08-14 4a22753
Update `CITATION.cff`
