+++ CITATION.cff
@@ -1,17 +1,16 @@
-cff-version: 1.2.0
-message: "If you use this source code, please cite it as below. Cite the journal article separately."
-title: superblockify: A Python Package for Automated Generation, Visualization, and Analysis of Potential Superblocks in Cities
-type: software
+cff-version: "1.2.0"
+title: "superblockify: A Python Package for Automated Generation,
+  Visualization, and Analysis of Potential Superblocks in Cities"
 authors:
-    - family-names: "Büth"
-      given-names: "Carlson Moses"
-      orcid: "https://orcid.org/0000-0003-2298-8438"
-    - family-names: "Vybornova"
-      given-names: "Anastassia"
-      orcid: "https://orcid.org/0000-0001-6915-2561"
-    - family-names: "Szell"
-      given-names: "Michael"
-      orcid: "https://orcid.org/0000-0003-3022-2483"
+- family-names: Büth
+  given-names: Carlson M.
+  orcid: "https://orcid.org/0000-0003-2298-8438"
+- family-names: Vybornova
+  given-names: Anastassia
+  orcid: "https://orcid.org/0000-0001-6915-2561"
+- family-names: Szell
+  given-names: Michael
+  orcid: "https://orcid.org/0000-0003-3022-2483"
 repository-code: 'https://github.com/NERDSITU/superblockify'
 url: 'https://superblockify.city/'
 keywords:
@@ -23,30 +22,30 @@ keywords:
   - urban mobility
   - urban data
 license: AGPL-3.0-or-later
-version: 1.0.0
 doi: 10.5281/zenodo.13300611
-date-released: 2024-08-12
-
-# add when article published in JOSS
-# preferred-citation:
-#   type: article
-#   authors:
-#     - family-names: "Büth"
-#       given-names: "Carlson Moses"
-#       orcid: "https://orcid.org/0000-0003-2298-8438"
-#     - family-names: "Vybornova"
-#       given-names: "Anastassia"
-#       orcid: "https://orcid.org/0000-0001-6915-2561"
-#     - family-names: "Szell"
-#       given-names: "Michael"
-#       orcid: "https://orcid.org/0000-0003-3022-2483"
-#   doi: "10.0000/00000"
-#   journal: "JOURNAL"
-#   month: 1
-#   start: 1
-#   end: 2
-#   title: "PROJECTNAME"
-#   issue: 1
-#   volume: 1
-#   year: YYYY
-#   abstract: "ABSTRACT"
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Büth
+    given-names: Carlson M.
+    orcid: "https://orcid.org/0000-0003-2298-8438"
+  - family-names: Vybornova
+    given-names: Anastassia
+    orcid: "https://orcid.org/0000-0001-6915-2561"
+  - family-names: Szell
+    given-names: Michael
+    orcid: "https://orcid.org/0000-0003-3022-2483"
+  date-published: 2024-08-14
+  doi: 10.21105/joss.06798
+  issn: 2475-9066
+  issue: 100
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 6798
+  title: "superblockify: A Python Package for Automated Generation,
+    Visualization, and Analysis of Potential Superblocks in Cities"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.06798"
+  volume: 9
