### 2024-08-14 #6798 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
